+++ bevy-trait-query/src/all/core/mod.rs
@@ -1,0 +1,3 @@
+pub(crate) mod fetch;
+pub(crate) mod read;
+pub(crate) mod write;
