+++ bevy-trait-query/src/all/impls/mod.rs
@@ -1,0 +1,3 @@
+mod all;
+
+pub use all::All;
