+++ bevy-trait-query/src/all/mod.rs
@@ -1,0 +1,5 @@
+mod core;
+mod impls;
+
+pub use core::{fetch::AllTraitsFetch, read::*, write::*};
+pub use impls::*;
