+++ bevy-trait-query/src/internal/mod.rs
@@ -1,0 +1,13 @@
+pub(crate) mod dyn_constructor;
+mod marker;
+mod register_ext;
+pub(crate) mod trait_registry;
+mod trait_state;
+mod zip_exact;
+
+pub use marker::*;
+pub use register_ext::*;
+pub use trait_state::*;
+
+pub(crate) use trait_registry::{TraitImplMeta, TraitImplRegistry};
+pub(crate) use zip_exact::zip_exact;
