+++ bevy-trait-query/src/one/core/mod.rs
@@ -1,0 +1,2 @@
+pub(crate) mod change_detection;
+pub(crate) mod fetch;
