+++ bevy-trait-query/src/one/mod.rs
@@ -1,0 +1,7 @@
+mod core;
+mod impls;
+
+pub use impls::*;
+
+pub use core::{change_detection::ChangeDetectionFetch, fetch::OneTraitFetch};
+pub(crate) use core::{change_detection::ChangeDetectionStorage, fetch::FetchStorage};
